+++ simdpp/detail/insn/i_reduce_and.h
@@ -41,7 +41,7 @@ SIMDPP_INL uint8_t i_reduce_and(const uint8x16& a)
 }
 
 #if SIMDPP_USE_AVX2
-SIMDPP_INL uint16_t i_reduce_and(const uint8x32& a)
+SIMDPP_INL uint8_t i_reduce_and(const uint8x32& a)
 {
     uint8x16 r = detail::extract128<0>(a);
     r = bit_and(r, detail::extract128<1>(a));
@@ -50,7 +50,7 @@ SIMDPP_INL uint16_t i_reduce_and(const uint8x32& a)
 #endif
 
 #if SIMDPP_USE_AVX512BW
-SIMDPP_INL uint16_t i_reduce_and(const uint8<64>& a)
+SIMDPP_INL uint8_t i_reduce_and(const uint8<64>& a)
 {
     uint8<32> r = detail::extract256<0>(a);
     r = bit_and(r, detail::extract256<1>(a));
